fix: moved throw error from DAL layer to BLL layer

+++ src/features/user-platform/application/usecases/delete-user.usecase.ts
@@ -1,5 +1,6 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { UsersRepository } from '../../repositories/users.repository';
+import { NotFoundException } from '@nestjs/common';
 
 export class DeleteUserCommand {
   constructor(public userId: string) {}
@@ -12,6 +13,11 @@ export class DeleteUserUseCase
   constructor(private usersRepository: UsersRepository) {}
 
   async execute(command: DeleteUserCommand): Promise<void> {
+    const user = await this.usersRepository.findById(command.userId);
+    if (!user) {
+      throw new NotFoundException();
+    }
+
     await this.usersRepository.deleteUser(command.userId);
   }
 }
